requirements.txt
Pillow>=10.0.0
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
python-vlc>=3.0.21203
